backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
twilio==9.2.3
gunicorn==21.2.0
